requirements.txt
fastapi>=0.110,<1
uvicorn[standard]>=0.29,<1
httpx>=0.27,<1
python-dotenv>=1.0,<2
pydantic>=2.12,<3
anyio>=4,<5
sniffio>=1.3,<2
requests
pydantic[email]